aliases/implications: factor out shared code.
