+++ app/models/tag_alias.rb
@@ -1,17 +1,8 @@
-class TagAlias < ApplicationRecord
-  attr_accessor :skip_secondary_validations
-
+class TagAlias < TagRelationship
   before_save :ensure_tags_exist
   after_save :clear_all_cache
   after_destroy :clear_all_cache
   after_save :create_mod_action
-  before_validation :initialize_creator, :on => :create
-  before_validation :normalize_names
-  validates_format_of :status, :with => /\A(active|deleted|pending|processing|queued|error: .*)\Z/
-  validates_presence_of :creator_id, :antecedent_name, :consequent_name
-  validates :creator, presence: { message: "must exist" }, if: lambda { creator_id.present? }
-  validates :approver, presence: { message: "must exist" }, if: lambda { approver_id.present? }
-  validates :forum_topic, presence: { message: "must exist" }, if: lambda { forum_topic_id.present? }
   validates_uniqueness_of :antecedent_name
   validate :absence_of_transitive_relation
   validate :antecedent_and_consequent_are_different
@@ -23,40 +14,6 @@ class TagAlias < ApplicationRecord
   belongs_to :forum_post
   attr_accessible :antecedent_name, :consequent_name, :forum_topic_id, :skip_secondary_validations
   attr_accessible :status, :approver_id, :as => [:admin]
-
-  module SearchMethods
-    def name_matches(name)
-      where("(antecedent_name like ? escape E'\\\\' or consequent_name like ? escape E'\\\\')", name.mb_chars.downcase.to_escaped_for_sql_like, name.downcase.to_escaped_for_sql_like)
-    end
-    
-    def active
-      where("status IN (?)", ["active", "processing"])
-    end
-
-    def search(params)
-      q = where("true")
-      return q if params.blank?
-
-      if params[:name_matches].present?
-        q = q.name_matches(params[:name_matches])
-      end
-
-      if params[:antecedent_name].present?
-        q = q.where("antecedent_name = ?", params[:antecedent_name])
-      end
-
-      if params[:id].present?
-        q = q.where("id in (?)", params[:id].split(",").map(&:to_i))
-      end
-
-      case params[:order]
-      when "created_at"
-        q = q.order("created_at desc")
-      end
-
-      q
-    end
-  end
 
   module CacheMethods
     extend ActiveSupport::Concern
@@ -124,7 +81,6 @@ class TagAlias < ApplicationRecord
     end
   end
 
-  extend SearchMethods
   include CacheMethods
   include ApprovalMethods
   include ForumMethods
@@ -170,32 +126,6 @@ class TagAlias < ApplicationRecord
         NewRelic::Agent.notice_error(e, :custom_params => {:tag_alias_id => id, :antecedent_name => antecedent_name, :consequent_name => consequent_name})
       end
     end
-  end
-
-  def is_pending?
-    status == "pending"
-  end
-
-  def is_active?
-    status == "active"
-  end
-  
-  def normalize_names
-    self.antecedent_name = antecedent_name.mb_chars.downcase.tr(" ", "_")
-    self.consequent_name = consequent_name.downcase.tr(" ", "_")
-  end
-
-  def initialize_creator
-    self.creator_id ||= CurrentUser.user.id
-    self.creator_ip_addr ||= CurrentUser.ip_addr
-  end
-
-  def antecedent_tag
-    Tag.find_or_create_by_name(antecedent_name)
-  end
-
-  def consequent_tag
-    Tag.find_or_create_by_name(consequent_name)
   end
 
   def absence_of_transitive_relation
@@ -308,17 +238,6 @@ class TagAlias < ApplicationRecord
         end
       end
     end
-  end
-
-  def deletable_by?(user)
-    return true if user.is_admin?
-    return true if is_pending? && user.is_builder?
-    return true if is_pending? && user.id == creator_id
-    return false
-  end
-
-  def editable_by?(user)
-    deletable_by?(user)
   end
 
   def reject!
